popovers: Move keyboard UI functions to modern module.
