+++ web/src/stream_popover.js
@@ -60,7 +60,7 @@ function get_popover_menu_items(sidebar_elem) {
 
 export function stream_sidebar_menu_handle_keyboard(key) {
     const items = get_popover_menu_items(stream_popover_instance);
-    popovers.popover_items_handle_keyboard(key, items);
+    popover_menus.popover_items_handle_keyboard(key, items);
 }
 
 export function elem_to_stream_id($elem) {
